requirements.txt
Django==4.2.16
djangorestframework==3.15.2
nltk==3.8.1
torch==2.1.2
transformers==4.43.3
